検索機能、ユーザー別表示設定

+++ resources/views/posts/show.blade.php
@@ -13,39 +13,52 @@
     
     <x-app-layout>
     <x-slot name="header">
-        　<p>レシピ一覧</p>
+        　<p>～レシピ詳細～</p>
     </x-slot>
     
     <body class="antialiased">
-            
+        <div class="pattern">    
         <div class='content'>
             
             <!--タイトル、写真、カテゴリー-->
             <div class="title">
                 <h1>『{{ $post->title }}』</h1>
+                <h3 class="category">カテゴリー：{{ $post->category->name }}</h2>
+            </div>
+            
+            <div class="user">
+                @if($post->user)
+                    <p class="user">作った人：{{ $post->user->name }}</p>
+                @endif
             </div>
             
             <div class="image">
-                <img src="{{ asset( $post->image) }}" alt="Post Image">
-            </div>
-            
-            <div class='category'>
-                <h2>カテゴリー：{{ $post->category->name }}</h2>
+                @if($post->image == null)
+                    <img src="{{ asset( "storage//jsI3pr0bNdUS1HypyALxx1uM7hHi2Sj6I6NltBUs.jpg" ) }}" alt="Post Image">
+                @else
+                    <img src="{{ asset( $post->image) }}" alt="Post Image">
+                @endif
             </div>
             
             <!--レシピ投稿内容-->
             <!--材料関連-->
-            <div class='content_post'>
+            <div class='content_post box25'>
                 <div class="ingredient">
-                    <h3>＜材料名＞</h3>
+                    <h3 class="ribbon1">材料</h3>
                         @foreach($ingredients as $ingredient) 
                             @if($ingredient->post_id === $post->id)
                                 <li class="content">
                                     @if($ingredient->name)
                                         {{ $ingredient->name }}
                                     @endif
+                                    @if($ingredient->front_unit)
+                                        {{ $ingredient->front_unit }}
+                                    @endif
                                     @if($ingredient->unit_amount)
                                         {{ $ingredient->unit_amount }}
+                                    @endif
+                                    @if($ingredient->back_unit)
+                                        {{ $ingredient->back_unit }}
                                     @endif
                                     @if($ingredient->unit_price)
                                         {{ $ingredient->unit_price }}円
@@ -54,15 +67,15 @@
                             @endif
                         @endforeach
                         
-                    <div class="sum_price">        
-                    <h3>＜合計金額＞</h3>
-                        <?php 
-                            $total_price = 0;
-                            foreach($ingredients as $ingredient)
-                                if($ingredient->post_id === $post->id && $ingredient->unit_price)
-                                    $total_price += $ingredient->unit_price;
-                        ?>
-                        {{ $total_price }}円
+                    <div class="sum_price"> 
+                            <?php 
+                                $total_price = 0;
+                                foreach($ingredients as $ingredient)
+                                    if($ingredient->post_id === $post->id && $ingredient->unit_price)
+                                        $total_price += $ingredient->unit_price;
+                            ?>
+                            <h2 class="sum">合計金額</h2>
+                            <p class="sum">{{ $total_price }}円</p>
                     </div>
                 </div>
                 
@@ -70,31 +83,42 @@
             </div>
             
             <!--作り方-->        
-            <div class='cook'>
-                <h2>＜作り方＞</h2>
-                <p　class='body'>『{{ $post->make}}』</p>
+            <div class='cook box25'>
+                <h2 class="ribbon1">作り方</h2>
+                <p class='body'>{!! nl2br(e($post->make)) !!}</p>
             </div>
             
-            <div class="reference">   
-                <h2>＜URL＞</h2>        
-                <h3>{{ $post->reference }}</h3>
+            <div class="reference box25">   
+                <h2 class="ribbon1">参考リンク</h2>        
+                <a href="{{ $post->reference }}"><h3 class="url">{{ $post->reference }}</h3></a>
             </div>
             
-        
-            <div class='edit'>
-                <a class="button" href="/posts/{{ $post->id }}/edit">変更する️</a>
-            </div>
-            
-            <form action="/posts/{{ $post->id }}" id="form_{{ $post->id }}" method="post">
-                @csrf
-                @method('DELETE')
-                <div class='delete'>
-                    <button class="button button_delete" type="button" onclick="deletePost({{ $post->id }})">削除する</button>
+            <div class="button-container">
+                <div class='post_week'>
+                    <a class="button return_button" href="/posts/{{ $post->id }}/menu">献立に追加️</a>
                 </div>
-            </form>
-            <div class='footer'>
-                <a href="/">◀戻る️</a>
+                
+                @if( $post->user->id == Auth::user()->id)
+                    <div class='edit'>
+                        <a class="button" href="/posts/{{ $post->id }}/edit">変更する️</a>
+                    </div>
+                    
+                    <form action="/posts/{{ $post->id }}" id="form_{{ $post->id }}" method="post">
+                        @csrf
+                        @method('DELETE')
+                        <div class='delete'>
+                            <button class="button button_delete" type="button" onclick="deletePost({{ $post->id }})">削除する</button>
+                        </div>
+                    </form>
+                @endif
+                
             </div>
+            
+            <div class='return'>
+                <a class="button return_button" href="/">◀戻る️</a>
+            </div>
+            
+        </div>
         </div>
         <script>
             function deletePost(id) {
